working version before layout adjustments

+++ controllers/recipes.js
@@ -2,6 +2,18 @@ const sha256 = require('js-sha256');
 
 module.exports = (db) => {
 
+    let isLoggedIn = function(request) {
+
+        if (request.cookies === undefined) {
+            return false;
+        }
+
+        const SALT = "not all lobsters are angry";
+        let loggedInHash = sha256(request.cookies['username'] + SALT);
+
+        return request.cookies['logged_in'] === loggedInHash;
+    }
+
     let allRecipes = (request, response) => {
         db.recipes.getAll((error, allRecipes) => {
             response.render('recipes/all', { allRecipes });
@@ -15,26 +27,23 @@ module.exports = (db) => {
     };
 
     let recipeForm = (request, response) => {
+
+        if (!isLoggedIn(request)) {
+            return response.redirect('/login');
+        }
         response.render('recipes/new')
     };
 
     let createRecipe = (request, response) => {
+
+        if (!isLoggedIn(request)) {
+            return response.redirect('/login');
+        }
         db.recipes.createOne(request, (error, recipe) => {
+            console.log(recipe);
             response.redirect(`recipes/${recipe[0].id}`);
         })
     };
-
-    let isLoggedIn = function(request) {
-
-        if (request.cookies === undefined) {
-            return false;
-        }
-
-        const SALT = "not all lobsters are angry";
-        let loggedInHash = sha256(request.cookies['username'] + SALT);
-
-        return request.cookies['logged_in'] === loggedInHash;
-    }
 
     let editForm = (request, response) => {
 
@@ -47,18 +56,30 @@ module.exports = (db) => {
     };
 
     let updateRecipe = (request, response) => {
+
+        if (!isLoggedIn(request)) {
+            return response.redirect('/login');
+        }
         db.recipes.updateOne(request, (error, recipe) => {
             response.redirect(`/recipes/${recipe[0].id}`);
         })
     };
 
     let deleteForm = (request, response) => {
+
+        if (!isLoggedIn(request)) {
+            return response.redirect('/login');
+        }
         db.recipes.getOne(request, (error, recipe) => {
             response.render('recipes/delete', { recipe });
         })
     };
 
     let deleteRecipe = (request, response) => {
+
+        if (!isLoggedIn(request)) {
+            return response.redirect('/login');
+        }
         db.recipes.deleteOne(request, (error, recipe) => {
             response.redirect('/recipes');
         })
